Add script to fix text errors and update .gitignore
The commit involves two changes. Firstly, a script is added under the “scripts” folder to automatically correct text errors in a 'global_original.ini' file, such as misspelt and incorrect forms, and put the corrected output in 'global_fixed.ini'. Secondly, the .gitignore file is updated to include 'venv', so that the virtual environment files are ignored by Git.

Signed-off-by: Lucas Greuloch (greluc) <lucas.greuloch@pm.me>