Simplify key checking process in key_test.py
Significant changes have been made to the key_test.py function to streamline and standardize the process of key checking. It now utilizes Python's configparser library to directly parse and compare the keys in two .ini files. Unnecessary operations and lines of code have been removed for efficiency and readability.

Signed-off-by: Lucas Greuloch (greluc) <lucas.greuloch@pm.me>