Enhance encoding test and add PTU files check
Expanded the functionality of the encoding_test.py script to include checks for both LIVE and PTU versions of the INI files. The check ensures that the encoding in the English and German versions of the INI files in both environments matches, and exits with error code if they differ.

Signed-off-by: Lucas Greuloch (greluc) <lucas.greuloch@pm.me>